pyproject.toml
[project]
name = "biov"
version = "0.1.1"
description = "Next-generation development experience for computational molecular biology."
readme = "README.md"
authors = [
    { name = "Tang Ziya", email = "tcztzy@gmail.Com" }
]
requires-python = ">=3.11"
dependencies = [
    "biopython>=1.85",
    "fsspec[full]>=2025.3.2",
    "pandas>=2.2.3",
    "pydantic-settings>=2.8.1",
    "typeguard>=4.4.2",
    "typer>=0.15.2",
    "typing-extensions>=4.13.2",
]

[project.scripts]
blat = "biov.executables:blat_app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/biov"]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    { include-group = "docs" },
    { include-group = "test" },
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.11",
    "mkdocstrings[python]>=0.29.1",
    "mkdocs-gen-files>=0.5",
    "mkdocs-literate-nav>=0.6.2",
]
test = [
    "pytest>=8.3.5",
]

[tool.ruff.lint]
preview = true
extend-select = ["D", "DOC", "I", "PD", "S", "T20"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
plugins = ['pydantic.mypy']

[tool.pyright]
venvPath = "."
venv = ".venv"
